# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KigaliWeShare API — neighbor-to-neighbor carpool matching for Kigali commuters
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listMyRideRequests
      tags: [requests]
      summary: Rides the current rider has requested
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [upcoming, past, all]
            default: upcoming
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RideRequest" }
    post:
      operationId: createRideRequest
      tags: [requests]
      summary: Rider requests a seat on a specific trip
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRideRequestBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RideRequest" }
  /requests/{requestId}/approve:
    post:
      operationId: approveRideRequest
      tags: [requests]
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RideRequest" }
  /requests/{requestId}/decline:
    post:
      operationId: declineRideRequest
      tags: [requests]
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RideRequest" }
  /requests/{requestId}/cancel:
    post:
      operationId: cancelRideRequest
      tags: [requests]
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RideRequest" }
...
components:
  schemas:
...
    Gender:
      type: string
      enum: [female, male, other, prefer_not_to_say]
...
    TripStatus:
      type: string
      enum: [scheduled, in_progress, completed, cancelled]
    RideRequestStatus:
      type: string
      enum: [pending, approved, declined, cancelled, completed]
...
    Trip:
      type: object
      required:
        [id, driverId, originId, destinationId, departureDate, departureTime, flexMinutes, seatsTotal, seatsRemaining, status]
      properties:
        id: { type: string }
        driverId: { type: string }
        driverName: { type: string }
        driverRating: { type: number, nullable: true }
        originId: { type: string }
        originName: { type: string }
        destinationId: { type: string }
        destinationName: { type: string }
        departureDate: { type: string, format: date }
        departureTime: { type: string, description: "HH:MM (window start)" }
        windowEndTime: { type: string, nullable: true, description: "HH:MM (window end)" }
        pickupPoint: { type: string, nullable: true }
        flexMinutes: { type: integer, enum: [5, 10, 15] }
        seatsTotal: { type: integer }
        seatsRemaining: { type: integer }
        sameGenderOnly: { type: boolean }
        notes: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/TripStatus" }
        cancelReason: { type: string, nullable: true }
        fuelShare:
          oneOf:
            - { $ref: "#/components/schemas/FuelShare" }
            - { type: "null" }
    FuelShare:
      type: object
      required:
        [distanceKm, pricePerLitreRwf, consumptionLPer100Km, fuelCostRwf, perPassengerRwf, driverPaysRwf, totalRiderShareRwf]
      description: |
        Suggested fuel cost-share for the trip. KigaliWeShare does not process payments;
        riders pay drivers directly (MoMo / cash). Per-rider amount can never exceed an
        equal split, ensuring this stays a non-commercial cost-share.
      properties:
        distanceKm: { type: number }
        pricePerLitreRwf: { type: number }
        consumptionLPer100Km: { type: number }
        fuelCostRwf: { type: number, description: "Total fuel cost for the whole trip" }
        perPassengerRwf: { type: number, description: "Suggested contribution per rider, rounded to 100 RWF" }
        driverPaysRwf: { type: number, description: "Driver's own out-of-pocket after rider contributions" }
        totalRiderShareRwf: { type: number }
...
    RideRequest:
      type: object
      required: [id, tripId, riderId, status, createdAt]
      properties:
        id: { type: string }
        tripId: { type: string }
        riderId: { type: string }
        riderName: { type: string }
        riderRating: { type: number, nullable: true }
        riderGender: { $ref: "#/components/schemas/Gender" }
        riderPhone: { type: string, nullable: true, description: "Revealed only after approval" }
        riderNeighborhood: { type: string, nullable: true }
        pickupPoint: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/RideRequestStatus" }
        createdAt: { type: string, format: date-time }
        trip:
          oneOf:
            - { $ref: "#/components/schemas/Trip" }
            - { type: "null" }
    CreateRideRequestBody:
      type: object
      required: [tripId]
      properties:
        tripId: { type: string }
        pickupPoint: { type: string, nullable: true }
        notes: { type: string, nullable: true }
...