# lib/api-zod/src/generated/types/createRideRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KigaliWeShare API — neighbor-to-neighbor carpool matching for Kigali commuters
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRideRequestBody {
  tripId: string;
  pickupPoint?: string | null;
  notes?: string | null;
}